worker/requirements.txt
# --- Core framework ---
fastapi==0.119.0
uvicorn==0.37.0
starlette==0.48.0

# --- AI / Strands integration ---
strands-agents==1.12.0
strands-agents-tools==0.2.11
google-genai==1.44.0
boto3==1.40.52
botocore==1.40.52

# --- HTTP / Utilities ---
requests==2.32.5
httpx==0.28.1
sse-starlette==3.0.2
aiohttp==3.13.0

# --- Environment & config ---
python-dotenv==1.1.1
django-environ==0.12.0
pydantic==2.12.0
pydantic-settings==2.11.0

# --- Optional (keep only if you use async speedups) ---
uvloop==0.21.0
anyio==4.11.0
sniffio==1.3.1
aio_pika